Refactor application and enable TFLite model
Refactor the application to condense the TFLite related logic to a
single class that initializes the interpreter and manages the inference
logic.

Allow the ability to pass the model and delegate path through the
command line for ease of testing with different models.

Remove quantization step and instead select the data from the cairo
surface directly based on the type used.

XNN models use float32 and Ethos models require quantized int8 models.